refactor: write each widget in a single file
