--- src/widgets/application_window/mod.rs
@@ -1,17 +1,0 @@
-mod imp;
-
-use adw::{gio, glib, Application};
-use glib::Object;
-
-glib::wrapper! {
-    pub struct ApplicationWindow(ObjectSubclass<imp::ApplicationWindow>)
-        @extends adw::ApplicationWindow, gtk::ApplicationWindow, gtk::Window, gtk::Widget,
-        @implements gio::ActionGroup, gio::ActionMap, gtk::Accessible, gtk::Buildable,
-                    gtk::ConstraintTarget, gtk::Native, gtk::Root, gtk::ShortcutManager;
-}
-
-impl ApplicationWindow {
-    pub fn new(app: &Application) -> Self {
-        Object::builder().property("application", app).build()
-    }
-}
